third_party/WebKit/LayoutTests/http/tests/security/isolatedWorld/bypass-main-world-csp-for-inline-script-expected.txt

ALERT: Injecting in main world: this should fail.
CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-eval'".

ALERT: Injecting into isolated world without bypass: this should fail.
CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-eval'".

ALERT: Starting to bypass main world's CSP: this should pass!
ALERT: EXECUTED in isolated world.
ALERT: Injecting into main world again: this should fail.
CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-eval'".

This test ensures that scripts run in isolated worlds marked with their own Content Security Policy aren't affected by the page's content security policy. Extensions, for example, should be able to inject inline JavaScript (even though it's probably a bad idea to do so).
